% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coprofile_example_data.R
\docType{data}
\name{coprofile_example_data}
\alias{coprofile_example_data}
\title{Example Data for Testing Spatial Epigenome-Transcriptome Co-Profiling Integration}
\format{
A named list with two elements:
\describe{
\item{ATAC}{A \code{data.frame} with metadata (first 4 columns) and normalized gene accessibility values.}
\item{RNA}{A \code{data.frame} with metadata (first 4 columns) and gene expression counts.}
}
Each data frame has the following structure:
\describe{
\item{Barcodes}{Unique spot barcodes for spatial transcriptomics (character).}
\item{x}{x-coordinates of spatial locations (numeric).}
\item{y}{y-coordinates of spatial locations (numeric).}
\item{Pathology.Annotations}{Pathology labels (factor with levels \code{"non-Corpus callosum"} and \code{"Corpus callosum"}).}
\item{Gabbr2, Pde7b, ..., Itga8}{Accessibility (ATAC) or expression (RNA) values for 100 selected genes (numeric).}
}
}
\usage{
data(coprofile_example_data)
}
\description{
This dataset contains a subset of processed spatial epigenome-transcriptome mouse brain
co-profiling data (Zhang et al., 2023), including 100 genes, intended for testing and
demonstrating the functionality of the package.
}
\examples{
data(coprofile_example_data)
}
\keyword{datasets}
