% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_wald_test.R
\name{run_gee_wald}
\alias{run_gee_wald}
\title{Perform GEE analysis with the robust Wald test for spatial transcriptomics}
\usage{
run_gee_wald(
  data,
  compare_levels,
  k = 100,
  family = poisson,
  corstr = "independence",
  cores = 1
)
}
\arguments{
\item{data}{A \code{data.frame} containing metadata in the first 4 columns and gene expression data in the remaining columns.
The metadata must include the following columns with these exact names:
\itemize{
\item \code{Barcodes}: Spot barcodes for spatial transcriptomics.
\item \code{x}: Spatial x-coordinates of the spots.
\item \code{y}: Spatial y-coordinates of the spots.
\item \code{Pathology.Annotations}: Pathology annotations for each spot.
The remaining columns should contain gene expression data.
}}

\item{compare_levels}{A character vector of length 2 specifying the levels of
\code{Pathology.Annotations} to compare. The first element of \code{compare_levels}
will be set as the reference level, and the second element will be set as
the comparison level in the factor.}

\item{k}{Number of clusters for k-means clustering (default: 100).}

\item{family}{A description of the error distribution and link function for the model (default: \code{poisson}).}

\item{corstr}{A character string specifying the correlation structure (default: \code{"independence"}).}

\item{cores}{Number of cores for parallel processing (default: 1).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{gene}: Gene names.
\item \code{p_value}: P-values from the robust Wald test for each gene.
}
}
\description{
This function performs Generalized Estimating Equations (GEE) analysis for each gene in the dataset,
comparing two user-specified levels of \code{Pathology.Annotations} using a robust Wald test.
}
\examples{
data(example_data)
results <- run_gee_wald(data = example_data, compare_levels = c("Fibrous Tissue", "Invasive Carcinoma"))
print(results)
}
