% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_independent.R
\name{run_indep_gee}
\alias{run_indep_gee}
\title{Perform GEE analysis by treating each spatial location as an independent observation for spatial transcriptomics}
\usage{
run_indep_gee(data, compare_levels, family = poisson, cores = 1)
}
\arguments{
\item{data}{A \code{data.frame} containing metadata in the first 4 columns and gene expression data in the remaining columns.
The metadata must include the following columns with these exact names:
\itemize{
\item \code{Barcodes}: Spot barcodes for spatial transcriptomics.
\item \code{x}: Spatial x-coordinates of the spots.
\item \code{y}: Spatial y-coordinates of the spots.
\item \code{Pathology.Annotations}: Pathology annotations for each spot.
The remaining columns should contain gene expression data.
}}

\item{compare_levels}{A character vector of length 2 specifying the levels of
\code{Pathology.Annotations} to compare. The first element of \code{compare_levels}
will be set as the reference level, and the second element will be set as
the comparison level in the factor.}

\item{family}{A description of the error distribution and link function for the model (default: \code{poisson}).}

\item{cores}{Number of cores for parallel processing (default: 1).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{gene}: Gene names.
\item \code{p_value}: P-values from the independent GEE analysis for each gene.
}
}
\description{
This function performs Generalized Estimating Equations (GEE) analysis for each gene in the dataset,
comparing two user-specified levels of \code{Pathology.Annotations} while treating each spatial location as independent.
The correlation structure is set to "independence," meaning no correlation is assumed between observations.
}
\examples{
data(example_data)
results <- run_indep_gee(data = example_data, compare_levels = c("Fibrous Tissue", "Invasive Carcinoma"))
print(results)
}
