% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{run_wilcoxon}
\alias{run_wilcoxon}
\title{Perform Wilcoxon rank-sum test for spatial transcriptomics}
\usage{
run_wilcoxon(data, compare_levels, cores = 1)
}
\arguments{
\item{data}{A \code{data.frame} containing metadata in the first 4 columns and gene expression data in the remaining columns.
The metadata must include the following columns with these exact names:
\itemize{
\item \code{Barcodes}: Spot barcodes for spatial transcriptomics.
\item \code{x}: Spatial x-coordinates of the spots.
\item \code{y}: Spatial y-coordinates of the spots.
\item \code{Pathology.Annotations}: Pathology annotations for each spot.
The remaining columns should contain gene expression data.
}}

\item{compare_levels}{A character vector of length 2 specifying the levels of
\code{Pathology.Annotations} to compare.}

\item{cores}{Number of cores for parallel processing (default: 1).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{gene}: Gene names.
\item \code{p_value}: P-values from the Wilcoxon rank-sum test for each gene.
}
}
\description{
This function performs the Wilcoxon rank-sum test for each gene in the dataset,
comparing two user-specified levels of \code{Pathology.Annotations}.
}
\examples{
data(example_data)
results <- run_wilcoxon(data = example_data, compare_levels = c("Fibrous Tissue", "Invasive Carcinoma"))
print(results)
}
